Feature: notificaciones, auditorias y mejoras - backend finalizado

+++ ms-clientes/src/main/java/core/cibertec/ms_clientes/application/usecase/DeleteClientUseCaseImpl.java
@@ -2,6 +2,7 @@ package core.cibertec.ms_clientes.application.usecase;
 
 import core.cibertec.ms_clientes.application.port.inservice.DeleteClientUseCase;
 import core.cibertec.ms_clientes.application.port.outservice.ClientRepositoryPort;
+import core.cibertec.ms_clientes.domain.exception.ClientNotFoundException;
 
 public class DeleteClientUseCaseImpl implements DeleteClientUseCase {
 
@@ -13,7 +14,9 @@ public class DeleteClientUseCaseImpl implements DeleteClientUseCase {
 
     @Override
     public void delete(Long id) {
-
+        if (repo.findById(id).isEmpty()) {
+            throw new ClientNotFoundException("Cliente no encontrado.");
+        }
         repo.deleteById(id);
     }
 }
